Changed flag system

+++ src/events/entityHitEntity/Crasher/crasher_b.ts
@@ -1,21 +1,20 @@
 import { Entity, Player, world } from '@minecraft/server';
 import config from '../../../data/config.js';
-import { flag, uniqueId, punish } from '../../../util/World.js';
+import { uniqueId } from '../../../util/World.js';
 import { State } from '../../../util/Toggle.js';
+import { flag } from '../../../util/Flag.js';
 
 const crasher_b = () => {
   const EVENT = world.afterEvents.entityHitEntity.subscribe(ev => {
     const player = ev.damagingEntity as Player;
-    if(player.typeId !== "minecraft:player") return;
+    if(player.typeId !== "minecraft:player" || uniqueId(player)) return;
     const hurtEntity: Entity = ev.damagingEntity;
     if(player.id == hurtEntity.id) {
       player.kill();
-      flag(player, 'crasher/B', 0);
-      punish(player, 'crasher/B', config.modules.crasherB.punishment)
-      if(!uniqueId(player)) punish(player, 'crasher/B', config.modules.crasherB.punishment)
+      flag(player, 'Crasher/B', config.modules.crasherB, [`hurtEntity=${hurtEntity.id}`])
     }
   });
-  if(!State('CRASHERB', config.modules.autoclickerA.state)) {
+  if(!State('CRASHERB', config.modules.crasherCB.state)) {
     world.afterEvents.entityHitEntity.unsubscribe(EVENT);
   };
 };
